Save - still fixing issues

+++ MyPoint.java
@@ -1,12 +1,12 @@
 public class MyPoint {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public MyPoint(int x, int y) {
+    public MyPoint(double x, double y) {
        setPoint(x, y);
     }
 
-    public void setPoint(int x, int y){
+    public void setPoint(double x, double y){
         this.x = x;
         this.y = y;
     }
